fix(product-components): remove circular dependency in SelectAssetModal
- Extracted AssetProps and AssetOptionBaseProps types to separate types.ts file
- AssetItem now imports types directly instead of from SelectAssetModal
- This breaks the circular import between SelectAssetModal and AssetItem
- Resolves: packages/product-components/src/components/SelectAssetModal/SelectAssetModal.tsx > packages/product-components/src/components/SelectAssetModal/AssetItem.tsx
