+++ packages/product-components/src/components/SelectAssetModal/SelectAssetModal.storiesData.ts
@@ -1,4 +1,4 @@
-import { AssetOptionBaseProps } from './SelectAssetModal';
+import type { AssetOptionBaseProps } from './types';
 
 interface SelectAssetOptionCurrencyProps extends AssetOptionBaseProps {
     type: 'currency';
