+++ packages/product-components/src/components/SelectAssetModal/SelectAssetModal.tsx
@@ -1,33 +1,12 @@
 import { ReactNode, useCallback, useState } from 'react';
 import { useIntl } from 'react-intl';
 
-import type { NetworkSymbolExtended } from '@suite-common/wallet-config';
-import { BaseCurrencyAmount } from '@suite-common/wallet-types';
 import { Column, Modal, VirtualizedList, useScrollShadow } from '@trezor/components';
 import { mapElevationToBackgroundToken, spacings } from '@trezor/theme';
 
 import { AssetItem } from './AssetItem';
 import { AssetItemNotFound } from './AssetItemNotFound';
-
-export interface AssetTokenBalance {
-    baseSymbol?: string; // TokenInfo['symbol'];
-    baseAmount: string;
-    fiatAmount: BaseCurrencyAmount | null;
-}
-
-export interface AssetProps {
-    ticker: string;
-    badge?: ReactNode;
-    symbol: NetworkSymbolExtended;
-    cryptoName?: string;
-    coingeckoId?: string;
-    contractAddress: string | null;
-    height: number;
-    shouldTryToFetch?: boolean;
-    tokenBalance?: AssetTokenBalance;
-}
-
-export type AssetOptionBaseProps = Omit<AssetProps, 'height' | 'formattedBalance'>;
+import type { AssetOptionBaseProps, AssetProps } from './types';
 
 export const ITEM_HEIGHT = 60;
 const LIST_MIN_HEIGHT = 200;
